flow analysis works correctly for all cardossier flows
