+++ graph-builder/src/main/java/com/github/lucacampanella/callgraphflows/staticanalyzer/instructions/InitiatingSubFlow.java
@@ -6,7 +6,7 @@ import com.github.lucacampanella.callgraphflows.graphics.components.GSubFlow;
 import com.github.lucacampanella.callgraphflows.graphics.components.GSubFlowWithCounterparty;
 import com.github.lucacampanella.callgraphflows.staticanalyzer.AnalysisResult;
 import com.github.lucacampanella.callgraphflows.staticanalyzer.Branch;
-
+import com.github.lucacampanella.callgraphflows.staticanalyzer.ClassDescriptionContainer;
 
 
 public class InitiatingSubFlow extends SubFlowBaseWithAnalysis {
@@ -54,9 +54,13 @@ public class InitiatingSubFlow extends SubFlowBaseWithAnalysis {
     public Branch getInstructionsForCombinations() {
         //we return just the flow if is an initiating flow
         //this will result on acceptCompanion being called on this instance
-        //for carda flows we'll just check if the flow matchesAnyChildren
+        //for corda flows we'll just check if the flow matchesAnyChildren
         //for initiating flow we'll get the other class and compare if they have at least one matching combination
         return new Branch(this);
+    }
+
+    public boolean containsValidProtocol() {
+        return resultOfClassAnalysis.containsValidProtocol();
     }
 }
