+++ graph-builder/src/main/java/com/github/lucacampanella/callgraphflows/staticanalyzer/instructions/Send.java
@@ -37,7 +37,7 @@ public class Send extends InstructionStatement implements StatementWithCompanion
         }
         send.sentType = analyzer.getCurrClassCallStackHolder().resolveEventualGenerics(
                 (((CtTypedElement) invocation.getArguments().get(0))).getType())
-                .box().getSimpleName();
+                .box().toString();
 
         send.targetSessionName = Optional.ofNullable(invocation.getTarget().toString());
