+++ graph-builder/src/main/java/com/github/lucacampanella/callgraphflows/staticanalyzer/instructions/StatementWithRelevantMethods.java
@@ -1,6 +1,7 @@
 package com.github.lucacampanella.callgraphflows.staticanalyzer.instructions;
 
 import com.github.lucacampanella.callgraphflows.staticanalyzer.AnalyzerWithModel;
+import com.github.lucacampanella.callgraphflows.staticanalyzer.Branch;
 import com.github.lucacampanella.callgraphflows.staticanalyzer.StaticAnalyzerUtils;
 import spoon.reflect.code.CtStatement;
 
@@ -21,5 +22,13 @@ public class StatementWithRelevantMethods extends InstructionStatement {
     @Override
     public boolean isRelevantForAnalysis() {
         return internalMethodInvocations.isRelevant();
+    }
+
+    @Override
+    public Branch desugar() {
+        Branch res = new Branch();
+        getInternalMethodInvocations().forEach(stmt -> res.addIfRelevantForAnalysis(stmt.desugar()));
+        //res.add(this); don't add this as it would result in a duplicate
+        return res;
     }
 }
